check if output exist to run

+++ run_hifi.bash
@@ -32,13 +32,17 @@ while IFS=',' read -r sample haplotype file_num datatype url; do
     hifi_bam_file="/nfs/turbo/dcmb-class/bioinf593/groups/group_05/raw/hifi/${sample}.hifi.bam"
     trgt_file="/nfs/turbo/dcmb-class/bioinf593/groups/group_05/output/trgt/${sample}.hifi.sorted.vcf.gz"
 
-    # Run each Snakefile with the specified configuration
-    snakemake -s "download.smk" -c "$config_file" --cores 1 --resources "mem_mb=1000"
-    snakemake -s "index.smk" -c "$config_file" --cores 1 --resources "mem_mb=4000"
-    snakemake -s "call_trgt.smk" -c "$config_file" --cores 1 --resources "mem_mb=32000"
+    # if trgt file does not exist, then run the pipeline
+    if [ ! -f "$trgt_file" ]; then
+        # Run each Snakefile with the specified configuration
+        snakemake -s "download.smk" -c "$config_file" --cores 1 --resources "mem_mb=1000"
+        snakemake -s "index.smk" -c "$config_file" --cores 1 --resources "mem_mb=4000"
+        snakemake -s "call_trgt.smk" -c "$config_file" --cores 1 --resources "mem_mb=32000"
 
-    # Delete hifi bam file and temp_manifest
-    rm "$hifi_bam_file"
+        # Delete hifi bam file and temp_manifest
+        rm "$hifi_bam_file"
+    fi
+
     rm "$temp_manifest"
     
 done < "$manifest_file"
